+++ tests/tests/test-02-mirtTwo.R
@@ -134,6 +134,11 @@ test_that('poly', {
                  tolerance = 1e-2)
     expect_equal(rsm@df, 9765610)
     expect_equal(grsm@df, 9765601)
+    graded <- mirt(data, 1, verbose = FALSE)
+    gM2 <- M2(graded, calcNull=TRUE)
+    expect_equal(gM2$M2, 6.52213, tolerance = 1e-4)
+    expect_equal(gM2$df.M2, 5)
+    expect_equal(gM2$CFI.M2, .9961693, tolerance = 1e-4)
 
     #item and test info
     Theta <- matrix(seq(-4,4,.01))